+++ packages/react-rtc/src/index.ts
@@ -1,0 +1,2 @@
+export { RtcProvider } from './RtcProvider';
+export * from './useRtc';
