+++ packages/react-rtc/src/types.ts
@@ -1,5 +1,10 @@
 export type Metadata = Record<string, unknown>;
 
+export type PeerConnection = Map<
+  string,
+  { displayName: string; pc: RTCPeerConnection; dataChannel: RTCDataChannel }
+>;
+
 export interface MessageData {
   message: string;
   id: string;
@@ -9,14 +14,25 @@ export interface MessageData {
   event?: Event;
   metadata?: Metadata;
 }
-export enum Event {
-  HAS_JOINED = 'hasJoined',
-  HAS_LEFT = 'hasLeft',
+
+export interface ContextType {
+  send: (inputValue: string) => void;
+  onEnter: (displayName: string, userMetadata?: Metadata) => void;
+  onLeave: () => void;
+  state: { isEntered: boolean };
+  messageData: MessageData[];
+  connections: PeerConnection;
+  error: string | null;
 }
 
 export interface User {
   displayName?: string;
   userMetadata?: Metadata;
+}
+
+export enum Event {
+  HAS_JOINED = 'hasJoined',
+  HAS_LEFT = 'hasLeft',
 }
 
 export enum ConnectionState {
