+++ packages/react-rtc/src/useRtc.ts
@@ -1,0 +1,4 @@
+import { useContext } from 'react';
+import { RtcContext } from './RtcContext';
+
+export const useRtc = () => useContext(RtcContext);
